+++ src/.internal/charts/hierarchy/HierarchyDefaultTheme.ts
@@ -45,6 +45,10 @@ export class HierarchyDefaultTheme extends Theme {
 			tooltipText: "{category}: {sum}"
 		});
 
+		r("HierarchyNode", ["last"]).setAll({			
+			cursorOverStyle: "default"
+		});		
+
 		{
 			const rule = r("Label", ["hierarchy", "node"]);